refactor: move `Circle` and `Column` to `modules/board/components`
